refactor: reorganize project structure by moving components and icons to common directories and cleaning up section files
