+++ src/components/pages/home/components/HeroContent.tsx
