+++ src/components/pages/home/components/LaptopMockup.tsx
@@ -3,7 +3,7 @@
 import { useLanguage } from "@/i18n/LanguageContext";
 import { useMockupLoader } from "./hooks/useMockupLoader";
 import { useLaptopZoom } from "./hooks/useLaptopZoom";
-import MockupLoader from "./MockupLoader";
+import MockupLoader from "./HeroMockups/MockupLoader";
 
 interface LaptopMockupProps {
   url?: string;
