+++ src/components/pages/legal/sections/LegalPageLayout.tsx
